+++ assets/src/design-system/utils/index.js
@@ -15,9 +15,10 @@
  */
 
 export { createContext, useContext, identity } from './context';
-export { default as useContextSelector } from './useContextSelector';
 export { default as isNullOrUndefinedOrEmptyString } from './isNullOrUndefinedOrEmptyString';
+export { noop } from './noop';
 export { default as useBatchingCallback } from './useBatchingCallback';
+export { default as useContextSelector } from './useContextSelector';
 export { default as useFocusOut } from './useFocusOut';
-export { default as useResizeEffect } from './useResizeEffect';
 export { default as addQueryArgs } from './addQueryArgs';
+export { default as useResizeEffect } from './useResizeEffect';
