+++ assets/src/design-system/utils/noop.js
